dex2oat tombstone with QCOM ART Optimization
Potentially the size of native stack of dex2oat is too small and
unfortunately it's unable to grow anymore because another chunk
of virtual space has been allocated right before the stack space,
in another word, there is no more room to grow the native stack.

A workaround is to grow the native stack to big enough after the
dex2oat is started.

[TheCrazyLex: Fixup definitions]

Change-Id: I3e8f9fe8677b109d7072172c1be4755487ad93bc
Signed-off-by: Alex Naidis <alex.naidis@linux.com>