DO NOT MERGE: Update `ArmInstructionSetFeatures` to track ARMv8-A.
Some instructions introduced in ARMv8-A will be useful to improve
the generated code.

Note that we remove mentions of cortex R and M cores from supported
variants. Android is only compiled for A class cores. Mentions of
cortex R and M class cores implies that the features detection should
work for them. This places additional constraints on the detection and
may prevent correct detection of some ARMv8-A features.

DO NOT MERGE: This will break the `FeaturesFromAssembly` test in
`runtime/arch/instruction_set_features_test.cc` on ARMv8-A CPUs.
Currently we compile with `-mcpu=cortex-a15` even for CPUs that
support ARMv8-A (eg. cortex-a53 and cortex-a57). So the
assembly-based feature detection will detect ARMv8-A while the feature
detection code based on preprocessor definitions will not.

Change-Id: I09ff7f5e8b41c872efe7d5581596040da7479d60