Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > > This is not needed. Just make sure you don't set
 > BOARD_USES_QCOM_HARDWARE
 > > in device tree. That flag has no other use in AOSPA.
 > 
 > Yup, I mentioned this in draft, I don't need to call this flag but
 > remove the "_arm" suffix.

Well, no device "needs"  it, as that's a proprietary module which doesn't ship in opensource and thus isn't used. 
The difference is that cm uses the qcom hardware flag for other purposes, in PA this flag is solely there for the additional performance modules. So setting it to true without having a BSP present would break build.

Patch-set: 2