compiler: image_writer: Hack to avoid persistent crashes
Change 8d26c59 prunes the dex cache when needed.

However the change also checks whether the prunning
happened for dex files which are in the class path.

This check completely crashes dex2oat on
some jar files.

This commit comments out the forced check
for prunning to avoid the dex2oat crash.

Since this is hacky, it should be reverted,
once a proper solution is in place.

Change-Id: Ic768c1cfc1f8412742ed248e66857151f42128be
Signed-off-by: Alex Naidis <alex.naidis@linux.com>