art: build: Increase optimization level
We can reach an improvement in cases
where SDCLANG is used to compile a
target with kryo or krait CPU,
by switching to Ofast.

This is safe for these targets,
since SDCLANG defines Ofast
differently from regular
clang.

Also we remove the IEEE 754 incompatible
options "fmath-errno" and "ffast-math"
to keep precision in floating point
calculations.

This applies only to target flags,
not to host flags.

Change-Id: Iafcc5a658a17f3ade600b1170c8bb66f0e775c26
Signed-off-by: Alex Naidis <alex.naidis@linux.com>