Update patch set 1
Patch Set 1:

This needs a bit of thinking, even if this is the target cpu (doesn't need a53 fixes per se), it can be and actually is bundled together with a53, so we need a way for it to actually use a53 workarounds. Right now it's in the list which indicates that this doesn’t need those.

Patch-set: 1
Label: Code-Review=0