dex2oat: Do not crash on InitializeClass() exceptions.
Previously `ClassLinker::EnsureInitialized()` could have
returned with a pending exception even when initializing
fields and parents was not allowed and the CompilerDriver
did not expect it, leading to a crash. Change the behavior
to consistently suppress exceptions when initializing
fields and/or parents is not allowed.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 184094466
Change-Id: I6a8af8c1da792c946f8f52ed4513ab1f0ccf2c9f
(cherry picked from commit ac57691c29ffa8a31d0b789492ee3f904c059f89)